requirements.txt
openai>=1.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
pydantic>=2.0.0 